refactor front page composition
