+++ src/app/pages/index.js
@@ -1,5 +1,6 @@
 import pipe from 'ramda/src/pipe';
 import _ArticlePage from './ArticlePage';
+import _FrontPage from './FrontPage';
 import _MediaAssetPage from './MediaAssetPage';
 import _PhotoGalleryPage from './PhotoGalleryPage';
 import _RadioPage from './RadioPage';
@@ -23,6 +24,7 @@ export const ArticlePage = pipe(
   applyBasicPageHandlers,
   withVariant,
 )(_ArticlePage);
+export const FrontPage = pipe(applyBasicPageHandlers, withVariant)(_FrontPage);
 export const MediaAssetPage = applyBasicPageHandlers(_MediaAssetPage);
 export const PhotoGalleryPage = applyBasicPageHandlers(_PhotoGalleryPage);
 export const RadioPage = applyBasicPageHandlers(_RadioPage);
